requirements/vps.txt
-r __init__.txt
django-less
gevent
gunicorn >= 0.15
